Update to 1.13.
Adds "verifyPortals" config option, which sets whether a stargate layout is verified when it is loaded. This can be used if, for example, the material ids in your previous gate layouts cannot match new material names, but you don't want existing stargates to break. Or if you want to change a stargate design with something like WorldEdit after it's built.