Fixing license
Stargate-Hmod did not have a stated license; Drakia chose GPL for Stargate-Bukkit.
In commit f898e1 Drakia decided to transition this project to LGPL.

For some reason, he duplicated the existing license and added the LGPL as `.lesser`.
That duplication creates unnecessary confusion and breaks github's licence preview functionality.

Given the above, LGPL is the correct license for this project; this commit reflects that.