A solution for issue #56
I kind of dislike this feature, and generally regard it as temporary (a temporary feature?).

How does it work? There's now another permission; stargate.enter which allows players to enter an already activated portal.

 Note that a portal closes when the ACTIVATOR enters it. So if one would want to enter a gate activated by another player, you would have to go through it before the activator (gate does not close when entered by someone not the activator)

If a player would have stargate.enter permission, but not have a permission set for a network, then the player would still have perms to enter. If in contrary the network permission is set, then stargate.enter permission follows according to that